+++ src/main/kotlin/org/codefreak/codefreak/service/AdministrationService.kt
@@ -2,6 +2,7 @@ package org.codefreak.codefreak.service
 
 import javax.servlet.MultipartConfigElement
 import org.codefreak.codefreak.Env
+import org.codefreak.codefreak.config.AppConfiguration
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.core.env.Environment
 import org.springframework.core.env.Profiles
@@ -11,14 +12,19 @@ import org.springframework.stereotype.Service
 class AdministrationService : BaseService() {
 
   @Autowired
-  lateinit var env: Environment
+  private lateinit var env: Environment
 
   @Autowired
-  lateinit var multipartConfigElement: MultipartConfigElement
+  private lateinit var config: AppConfiguration
+
+  @Autowired
+  private lateinit var multipartConfigElement: MultipartConfigElement
 
   fun getMotd() = if (env.acceptsProfiles(Profiles.of(Env.DEMO)))
       "This is a demo system. Data may disappear at any time!"
       else null
 
   fun getMaxUploadSize(): Long = multipartConfigElement.maxFileSize
+
+  fun getDefaultIdeDockerImageName(): String = config.ide.image
 }